refactor : 기존의 체스 기물들을 리팩토링 된 체스 기물로 변경
