+++ src/main/java/chess/controller/ChessController.java
@@ -1,5 +1,6 @@
 package chess.controller;
 
+
 import chess.domain.board.Position;
 import chess.domain.game.ChessGame;
 import chess.view.InputView;
