+++ src/main/java/chess/domain/piece/King.java
@@ -1,18 +1,22 @@
 package chess.domain.piece;
 
-import chess.domain.board.Position;
+import static chess.domain.board.Direction.DOWN;
+import static chess.domain.board.Direction.DOWN_LEFT;
+import static chess.domain.board.Direction.DOWN_RIGHT;
+import static chess.domain.board.Direction.LEFT;
+import static chess.domain.board.Direction.RIGHT;
+import static chess.domain.board.Direction.TOP;
+import static chess.domain.board.Direction.TOP_LEFT;
+import static chess.domain.board.Direction.TOP_RIGHT;
+
 import chess.domain.piece.attribute.Color;
-import chess.domain.piece.attribute.Name;
-import chess.domain.piece.strategy.KingMoveStrategy;
+import java.util.Arrays;
 
-public class King extends Piece {
+public class King extends FixedMovablePiece {
+
     public King(Color color) {
-        super(new Name("K"), color, new KingMoveStrategy());
-    }
-
-    @Override
-    public boolean canMove(Piece targetPiece, Position from, Position to) {
-        return moveStrategy.isValidateCanMove(color, from, to);
+        super(color, "K", Arrays.asList(
+                TOP, DOWN, RIGHT, LEFT, TOP_RIGHT, TOP_LEFT, DOWN_RIGHT, DOWN_LEFT));
     }
 
     @Override
