+++ src/main/java/chess/domain/piece/MultipleMovablePiece.java
@@ -1,9 +1,9 @@
-package chess.domain.refactorPiece;
+package chess.domain.piece;
 
+import chess.domain.board.ChessBoard;
+import chess.domain.board.Position;
 import chess.domain.piece.attribute.Color;
-import chess.domain.refactorBoard.ChessBoard;
-import chess.domain.refactorBoard.Direction;
-import chess.domain.refactorPosition.Position;
+import chess.domain.board.Direction;
 import java.util.List;
 
 public abstract class MultipleMovablePiece extends Piece {
