--- src/main/java/chess/domain/piece/attribute/Name.java
@@ -1,32 +1,0 @@
-package chess.domain.piece.attribute;
-
-import java.util.Objects;
-
-public class Name {
-    private final String value;
-
-    public Name(String value) {
-        this.value = value;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof Name)) {
-            return false;
-        }
-        Name name = (Name) o;
-        return Objects.equals(value, name.value);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(value);
-    }
-}
