+++ src/test/java/chess/domain/board/DirectionTest.java
@@ -2,10 +2,8 @@ package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.piece.attribute.Color;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
@@ -13,63 +11,19 @@ class DirectionTest {
 
     @ParameterizedTest
     @CsvSource(value = {
-            "A,TWO,A,THREE,TOP",
-            "A,SEVEN,A,SIX,DOWN",
-            "A,TWO,B,TWO,RIGHT",
-            "B,TWO,A,TWO,LEFT",
-            "A,TWO,B,THREE,TOPRIGHT",
-            "B,TWO,A,THREE,TOPLEFT",
-            "A,SEVEN,B,SIX,DOWNRIGHT",
-            "B,SEVEN,A,SIX,DOWNLEFT",
-            "B,ONE,C,THREE,TTR",
-            "B,ONE,D,TWO,RRT",
-            "B,TWO,D,ONE,RRD",
-            "B,THREE,C,ONE,DDR",
-            "B,THREE,A,ONE,DDL",
-            "C,TWO,A,ONE,LLD",
-            "C,TWO,A,THREE,LLT",
-            "B,ONE,A,THREE,TTL"
-    })
-    @DisplayName("체스 말이 거리가 같은지 확인할 수 있다.")
-    void isSameDistance(File fileA, Rank rankA, File fileB, Rank rankB, Direction direction) {
+            "a,2,a,2,TOP,0", "a,2,a,3,TOP,1", "a,2,a,4,TOP,2",
+            "a,7,a,7,DOWN,0", "a,7,a,6,DOWN,1", "a,7,a,5,DOWN,2",
+            "a,3,a,3,RIGHT,0", "a,3,h,3,RIGHT,7",
+            "h,3,h,3,LEFT,0", "h,3,a,3,LEFT,7",
+            "a,2,b,3,TOP_RIGHT,1", "a,2,f,7,TOP_RIGHT,5",
+            "f,2,e,3,TOP_LEFT,1", "f,2,a,7,TOP_LEFT,5",
+            "a,7,b,6,DOWN_RIGHT,1", "a,7,f,2,DOWN_RIGHT,5",
+            "f,7,e,6,DOWN_LEFT,1", "f,7,a,2,DOWN_LEFT,5"})
+    @DisplayName("기물의 현재 위치에서 최종 위치로 이동할 때 이동하는 모든 위치 값들을 구할 수 있다.")
+    void findByRoute(String fromColumn, String fromRow, String toColumn, String toRow, Direction direction, int size) {
+        final List<Position> directions = direction.route(
+                Position.of(fromColumn, fromRow), Position.of(toColumn, toRow));
 
-        assertThat(direction.isSameDistance(
-                new Position(fileA, rankA),
-                new Position(fileB, rankB)))
-                .isTrue();
-    }
-
-    @Test
-    @DisplayName("위치 두개로 방향 값을 가져올 수 있다.")
-    void getDirection() {
-        assertThat(Direction.of(new Position(File.A, Rank.ONE),
-                new Position(File.C, Rank.THREE)))
-                .isEqualTo(Direction.TOPRIGHT);
-
-    }
-
-    @Test
-    @DisplayName("진영에 따라서 말의 이동방향이 뒤집힌다.")
-    void pieceDirectionReversed() {
-        List<Direction> directions = Direction.pawnDirection(Color.BLACK);
-
-        assertThat(directions).containsAll(
-                List.of(Direction.DOWN, Direction.DOWNLEFT, Direction.DOWNRIGHT)
-        );
-    }
-
-    @ParameterizedTest
-    @CsvSource(value = {
-            "A,TWO,A,THREE,TOP",
-            "A,SEVEN,A,SIX,DOWN",
-            "A,TWO,B,TWO,RIGHT",
-            "B,TWO,A,TWO,LEFT"
-    })
-    @DisplayName("체스 말이 해당 방향으로 가는지 확인할 수 있다.")
-    void isSameDirection(File fileA, Rank rankA, File fileB, Rank rankB, Direction direction) {
-        assertThat(direction.isSameDirection(
-                new Position(fileA, rankA),
-                new Position(fileB, rankB)))
-                .isTrue();
+        assertThat(directions).hasSize(size);
     }
 }
